refactor : 시급입력.  register.mypage 모두에서 사용할 수 있는 WageForm 컴포넌트 작성
