+++ src/app.features/mypage/screens/WageSettingScreen.tsx
@@ -1,8 +1,7 @@
-import { useRouter } from 'next/router';
-import React, { useEffect, useState } from 'react';
-import TextInput from 'src/app.components/app.base/Input/TextInput';
+import React, { useState } from 'react';
 import InputInteractButton from 'src/app.components/Button/InputInteractButton';
 import Header from 'src/app.components/Header';
+import WageForm from 'src/app.components/UserForm/WageForm';
 import { MutateTpye } from 'src/app.modules/api/client';
 import { MutateUserBodyType } from 'src/app.modules/api/user';
 import { IUser } from 'src/app.modules/types/user';
@@ -14,7 +13,7 @@ interface Props {
 }
 function WageSettingScreen({ user, putUser, isLoading }: Props) {
 	const [wage, setWage] = useState<string>(`${user?.wage ?? ''}`);
-	const submitHandler = () => {
+	const submitHandler = (): void => {
 		if (isLoading) return;
 		if (!wage) return;
 		const body = {
@@ -24,27 +23,13 @@ function WageSettingScreen({ user, putUser, isLoading }: Props) {
 
 		putUser(body);
 	};
-	const wageHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-		const newWage = e.target.value;
-		if (!Number(newWage) && newWage.trim() !== '') return;
-		setWage(newWage);
-	};
-	const resetwageHandler = () => {
-		setWage('');
-	};
 
 	return (
 		<>
 			<Header title="시급 수정" />
 			<main className=" pt-[7.2rem] space-y-[0.8rem]">
 				<h2 className="text-g6 text-subhead1">시급입력</h2>
-				<TextInput
-					value={wage ?? ''}
-					onChange={wageHandler}
-					resetHandler={resetwageHandler}
-					mode="default"
-					placeholder="9,620 원"
-				/>
+				<WageForm wage={`${wage ?? ''}`} onWageChange={setWage} />
 				<InputInteractButton disabled={!wage} onClick={submitHandler} />
 			</main>
 		</>
